requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
tweepy>=4.14.0
praw>=7.7.1
python-telegram-bot>=20.7

# AI and Data Processing
numpy>=1.24.0
sqlite3  # Built-in with Python

# Timezone and Date Handling
pytz>=2023.3

# Async Support
asyncio  # Built-in with Python
asyncpraw>=7.7.0  # For async Reddit operations (optional)

# HTTP and Connection Management
httpx>=0.25.0

# Logging and Monitoring
logging  # Built-in with Python

# Data Serialization
json  # Built-in with Python

# Cryptographic Functions
hashlib  # Built-in with Python

# UUID Generation
uuid  # Built-in with Python

# Threading Support
threading  # Built-in with Python

# Random Number Generation
random  # Built-in with Python

# Operating System Interface
os  # Built-in with Python

# Time Utilities
time  # Built-in with Python
datetime  # Built-in with Python

# Type Hints
typing  # Built-in with Python

# Collections
collections  # Built-in with Python

# Optional: Enhanced Analytics
# pandas>=2.0.0  # Uncomment for advanced data analysis
# matplotlib>=3.7.0  # Uncomment for data visualization
# seaborn>=0.12.0  # Uncomment for statistical plots

# Optional: Machine Learning (Future Enhancements)
# scikit-learn>=1.3.0  # Uncomment for ML features
# tensorflow>=2.13.0  # Uncomment for deep learning
# torch>=2.0.0  # Uncomment for PyTorch support

# Development Dependencies (Optional)
# pytest>=7.4.0  # For testing
# black>=23.0.0  # For code formatting
# flake8>=6.0.0  # For linting
# mypy>=1.5.0  # For type checking
